refactor: 서블릿 패키지명 및 구조 수정
