+++ tomcat/src/main/java/nextstep/jwp/controller/UserController.java
@@ -1,10 +1,10 @@
 package nextstep.jwp.controller;
 
-import static org.apache.catalina.servlet.handler.ServletResponse.redirect;
+import static org.apache.catalina.core.servlet.ServletResponse.redirect;
 
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
-import org.apache.catalina.servlet.handler.ServletResponse;
+import org.apache.catalina.core.servlet.ServletResponse;
 import org.apache.coyote.http11.common.Cookies;
 import org.apache.coyote.http11.common.QueryParameters;
 import org.apache.coyote.http11.request.Request;
