+++ tomcat/src/main/java/org/apache/catalina/connector/Connector.java
@@ -8,7 +8,7 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.LinkedBlockingQueue;
 import java.util.concurrent.ThreadPoolExecutor;
 import java.util.concurrent.TimeUnit;
-import org.apache.catalina.servlet.RequestHandlerAdaptor;
+import org.apache.catalina.core.RequestHandlerAdaptor;
 import org.apache.coyote.http11.Http11Processor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
