+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -1,8 +1,8 @@
 package org.apache.coyote.http11;
 
 import java.net.Socket;
-import org.apache.catalina.servlet.RequestHandlerAdaptor;
-import org.apache.catalina.servlet.handler.ServletResponse;
+import org.apache.catalina.core.RequestHandlerAdaptor;
+import org.apache.catalina.core.servlet.ServletResponse;
 import org.apache.coyote.http11.io.RequestReader;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
