+++ tomcat/src/main/java/org/apache/coyote/http11/common/QueryParameters.java
@@ -3,7 +3,7 @@ package org.apache.coyote.http11.common;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.apache.catalina.servlet.util.StringFormatParser;
+import org.apache.catalina.core.util.StringFormatParser;
 
 public class QueryParameters {
