+++ tomcat/src/test/java/nextstep/jwp/controller/UserControllerTest.java
@@ -6,7 +6,7 @@ import static org.apache.coyote.http11.common.header.HeaderName.COOKIE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
-import org.apache.catalina.servlet.handler.ServletResponse;
+import org.apache.catalina.core.servlet.ServletResponse;
 import org.apache.coyote.http11.common.Cookies;
 import org.apache.coyote.http11.common.Status;
 import org.apache.coyote.http11.request.Request;
