+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/DispatcherServletFixture.java
@@ -3,7 +3,7 @@ package nextstep.org.apache.coyote.http11;
 import java.util.Set;
 import nextstep.jwp.handler.LoginRequestHandler;
 import nextstep.jwp.handler.RegisterRequestHandler;
-import org.apache.catalina.servlet.RequestHandlerAdaptor;
+import org.apache.catalina.core.RequestHandlerAdaptor;
 
 public class DispatcherServletFixture {
