+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -8,7 +8,7 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import javassist.NotFoundException;
-import org.apache.catalina.servlet.util.ResourceReader;
+import org.apache.catalina.core.util.ResourceReader;
 import org.apache.coyote.http11.Http11Processor;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
