+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/RequestHandlerAdaptorTest.java
@@ -8,7 +8,7 @@ import static org.apache.coyote.http11.common.header.HeaderName.ACCEPT;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
-import org.apache.catalina.servlet.handler.ServletResponse;
+import org.apache.catalina.core.servlet.ServletResponse;
 import org.apache.coyote.http11.common.Status;
 import org.apache.coyote.http11.request.Request;
 import org.apache.coyote.http11.response.Response;
